webapi/tct-backgrounds-css3-tests/backgrounds/CSS3BG_border-top-left-radius_length_length.html

<!DOCTYPE html>
<!--
Copyright (c) 2012 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Tan, shiyou <shiyoux.tan@intel.com>

-->

<html>
  <head>
    <title>CSS3 Backgrounds Test: CSS3BG_border-top-left-radius_length_length</title>
    <link rel="author" title="Intel" href="http://www.intel.com/" />
    <link rel="help" href="http://www.w3.org/TR/2012/CR-css3-background-20120724/#border-radius" />
    <meta name="assert" content="Check if the element border-top-left-radius test expected '{1px, 4px}'" />
    <meta name="flags" content="" />
    <script src="support/support.js"></script>
    <script src="../resources/testharness.js" type="text/javascript"></script>
    <script src="../resources/testharnessreport.js" type="text/javascript"></script>
  </head>
  <body>
    <div id="log"></div>
    <div id="test"></div>
    <script>
        test(function(){
             var div = document.querySelector("#test");
             div.style[headProp("border-top-left-radius")] = "1px 4px";
             var border = GetCurrentStyle("border-top-left-radius");
             assert_equals(border, "1px 4px", "border-top-left-radius value Test");
        }, document.title);
    </script>
  </body>
</html>
